T8511 De-duplicate code in `/aukojimas`
