+++ src/pages/kaip-galiu-padeti/aukojimas/uzsienyje.js
@@ -1,26 +1,10 @@
 import * as React from "react";
 import PropTypes from "prop-types";
 import { graphql } from "gatsby";
-import { GatsbyImage, getImage, StaticImage } from "gatsby-plugin-image";
-import { Title, Meta } from "react-head";
 
-// Styles.
-import "./aukojimas.css";
-
-// Components.
-import Card from "../../../components/Card";
-import CardList from "../../../components/CardList";
-import Constraint from "../../../components/Constraint";
-import Layout from "../../../components/Layout";
-import NavigationGroup from "../../../components/NavigationGroup";
-import Button from "../../../components/Button";
-import CardSection from "../../../components/Card/CardSection";
-import Section from "../../../components/Section";
-import TabsButton from "../../../components/TabsButton";
+import Aukojimas from "../../../shared-pages/Aukojimas";
 
 const Page = ({ data }) => {
-  const crumbs = [`Aukojimas`];
-  const additionalNavigation = [`Savanorystė`];
   const content = data.contents.edges.map((edge) => {
     return {
       ...edge.node.childMarkdownRemark.frontmatter,
@@ -34,94 +18,11 @@ const Page = ({ data }) => {
   });
 
   return (
-    <Layout pagePath="/kaip-galiu-padeti/aukojimas/uzsienyje/">
-      <Title>Aukojimas</Title>
-
-      <Section className="HeroSectionB">
-        <StaticImage
-          className="HeroSectionB__background"
-          src="../../../images/hero/aukojimas.jpg"
-          alt="Su Ukraina!"
-          layout="fullWidth"
-        />
-      </Section>
-
-      {!!content && (
-        <Constraint>
-          <NavigationGroup
-            crumbs={crumbs}
-            additionalNav={additionalNavigation}
-          />
-          <h1>{content.title}</h1>
-          <div dangerouslySetInnerHTML={{ __html: content.html }} />
-          <Meta name="description" content={content.excerpt} />
-        </Constraint>
-      )}
-
-      <Constraint>
-        <section className="TabsContainer">
-          <TabsButton
-            to={`/kaip-galiu-padeti/aukojimas/lietuvoje/`}
-            text={`Lietuvoje`}
-          />
-          <TabsButton
-            to={`/kaip-galiu-padeti/aukojimas/uzsienyje/`}
-            text={`Užsienyje`}
-            active={`active`}
-          />
-        </section>
-        <CardList>
-          {organisations.map((organisation, i) => {
-            const logo = organisation.logo && (
-              <GatsbyImage
-                image={getImage(organisation.logo)}
-                alt={organisation.title}
-              />
-            );
-            return (
-              <Card title={organisation.title} logo={logo} key={i}>
-                {!!organisation.about && (
-                  <CardSection title="Apie" content={organisation.about} />
-                )}
-                {!!organisation.cause && (
-                  <CardSection title="Paskirtis" content={organisation.cause} />
-                )}
-                {!!organisation.rekvizitai && (
-                  <CardSection
-                    title="Kita informacija"
-                    content={organisation.rekvizitai}
-                  />
-                )}
-
-                <div className={`Card__actions`}>
-                  {!!organisation.support_link && (
-                    <Button
-                      icon={`arrow-white`}
-                      href={organisation.support_link}
-                      color={`primary`}
-                      text={`Paremti`}
-                      position={`right`}
-                      target="_blank"
-                    />
-                  )}
-
-                  {!!organisation.website && (
-                    <Button
-                      icon={`arrow-blue`}
-                      href={organisation.website}
-                      color={`transparent`}
-                      text={`Oficialus puslapis`}
-                      position={`right`}
-                      target="_blank"
-                    />
-                  )}
-                </div>
-              </Card>
-            );
-          })}
-        </CardList>
-      </Constraint>
-    </Layout>
+    <Aukojimas
+      content={content}
+      organisations={organisations}
+      pagePath="/kaip-galiu-padeti/aukojimas/uzsienyje/"
+    />
   );
 };
 
@@ -185,9 +86,7 @@ Page.propTypes = {
         PropTypes.shape({
           node: PropTypes.shape({
             childMarkdownRemark: PropTypes.shape({
-              frontmatter: PropTypes.shape({
-                title: PropTypes.string,
-              }),
+              frontmatter: PropTypes.object,
             }),
             html: PropTypes.string,
             excerpt: PropTypes.string,
@@ -200,16 +99,7 @@ Page.propTypes = {
         PropTypes.shape({
           node: PropTypes.shape({
             childMarkdownRemark: PropTypes.shape({
-              frontmatter: PropTypes.shape({
-                about: PropTypes.string,
-                cause: PropTypes.string,
-                rekvizitai: PropTypes.string,
-                title: PropTypes.string,
-                website: PropTypes.string,
-                support_link: PropTypes.string,
-                logo: PropTypes.object,
-                location: PropTypes.string,
-              }),
+              frontmatter: PropTypes.object,
             }),
           }),
         })
